Trying to fix paging errors, Added vscode debug configs

+++ kernel/arch/x86_64/mmu/mmu.c
@@ -4,5 +4,5 @@ uint64_t aalign(uint64_t addr, uint64_t alignment) {
     // Align addr by alignment
     if (addr % alignment == 0)
         return addr;
-    return ((addr / alignment + 1) * alignment);
+    return (((int) (addr / alignment + 1)) * alignment);
 }
